split AboutUs section into 3 separate components and move styles
