+++ courses/frontend/react/space-galactica-app/src/pages/AboutUsPage/AboutUsPage.module.css
@@ -3,3 +3,77 @@
   width: 100vw;
   height: 100vh;
 }
+
+/*--ourValue--*/
+.valuesGrid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
+  gap: 20px;
+}
+
+.valueCard {
+  background: rgba(255, 255, 255, 0.05);
+  padding: 20px;
+  border-radius: 10px;
+}
+
+/*--ourCrew--*/
+.crewSection {
+  background-image: url("/crew/background-crew-desktop.jpg");
+  background-size: cover;
+  background-position: center;
+  color: #fff;
+  position: relative;
+}
+
+.crewSection::before {
+  content: "";
+  position: absolute;
+  inset: 0;
+  background: rgba(0, 0, 0, 0.3);
+}
+
+.content {
+  z-index: 1;
+  position: relative;
+  padding: 15px;
+}
+
+p {
+  text-align: justify;
+}
+
+.content ul li {
+  list-style-type: none;
+  margin: 30px 10px;
+}
+
+.content ul li strong {
+  display: inline-block;
+  margin-bottom: 10px;
+  margin-left: -10px;
+}
+
+/*--ourPartner--*/
+.partnerGrid {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 20px;
+  align-items: center;
+  justify-items: center;
+  margin-top: 30px;
+}
+
+.partnerGrid img {
+  width: 120px;
+}
+
+@media (max-width:650px) {
+  .partnerGrid {
+    grid-template-columns: repeat(2, 1fr);
+  }
+
+  .partnerGrid img {
+    margin: 15px;
+  }
+}
